requirements.txt
asn1crypto==0.24.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
praw==6.4.0
prawcore==1.0.1
pycparser==2.19
python-telegram-bot==12.1.1
requests==2.22.0
six==1.12.0
tornado==6.0.3
update-checker==0.16
urllib3==1.25.6
websocket-client==0.56.0
